simplify stack trace calculation
